Move the SecurityContectHolderStrategy bean from VaadinWebSecurity (#15059)
